+++ bin/miden-cli/src/info.rs
@@ -33,10 +33,8 @@ async fn print_client_stats(client: &Client<impl FeltRng>) -> Result<(), String>
 }
 
 fn print_config_stats(config: &CliConfig) -> Result<(), String> {
-    println!("Node address: {}", config.rpc.endpoint.host());
-    let store_len = fs::metadata(config.store.database_filepath.clone())
-        .map_err(|e| e.to_string())?
-        .len();
+    println!("Node address: {}", config.rpc.endpoint.0.host());
+    let store_len = fs::metadata(config.store_filepath.clone()).map_err(|e| e.to_string())?.len();
     println!("Store size: {} kB", store_len / 1024);
     println!(
         "Default account: {}",