+++ crates/rust-client/src/rpc/mod.rs
@@ -11,6 +11,7 @@ use domain::{
     notes::{NoteDetails, NoteSyncInfo},
     sync::StateSyncInfo,
 };
+
 use miden_objects::{
     accounts::AccountId,
     crypto::merkle::MmrProof,