+++ crates/rust-client/src/rpc/tonic_client/mod.rs
@@ -38,6 +38,7 @@ use crate::{
     rpc::{generated::requests::GetBlockHeaderByNumberRequest, RpcError},
 };
 
+
 // TONIC RPC CLIENT
 // ================================================================================================
 
@@ -51,12 +52,13 @@ pub struct TonicRpcClient {
 }
 
 impl TonicRpcClient {
-    /// Returns a new instance of [TonicRpcClient] that'll do calls the `config_endpoint` provided
-    pub fn new(config: &RpcConfig) -> TonicRpcClient {
+    /// Returns a new instance of [TonicRpcClient] that'll do calls to the provided [Endpoint] with
+    /// the given timeout in milliseconds
+    pub fn new(endpoint: Endpoint, timeout_ms: u64) -> TonicRpcClient {
         TonicRpcClient {
             rpc_api: None,
-            endpoint: config.endpoint.to_string(),
-            timeout_ms: config.timeout_ms,
+            endpoint: endpoint.to_string(),
+            timeout_ms,
         }
     }